Update BMSDao.sol
To Start there will be tiers of release to the DAO to assure that roadmapped components can reach fruition, and find a stable presence.   

Initial elements that can be governed by early DAO governance include,but are not limited to:
Proposals and Voting on which tokens to offer in games.
Proposals and Voting on algorithmic strategies for determining the Moon Price in the Benevolent Moon Squares Game.
Proposals and Voting on the number of tokens to be offered to play. 
Proposals and Voting on Benevolent Organizations to offer users. 
Proposals and Voting on which tokens to accept for entry fees.


Secondary areas of governance upon the launch and stability of the NFT lottery and galleries include but are not limited to: 
Proposals and Voting on method and process of executing the NFT Lottery
Proposals and Voting on determining  BMS fees and use for NFT trading
Proposals and Voting on Benevolent Organizations to receive percentages of trading fees . 

Tertiary areas of governance upon launch of liquidity, lending and borrowing instruments (Defi). 
Proposals and Voting on which initial pairs to offer (along with BMS)
Proposals and Voting on which Benevolent Organizations will receive percentage of trading fees. 
Proposals and Voting on algorithmic process for assigning lending and borrowing interest rates.